refactor: Prevent import from going the wrong way
