+++ plugins/serializers/wavefront/wavefront_test.go
@@ -11,7 +11,6 @@ import (
 
 	"github.com/influxdata/telegraf"
 	"github.com/influxdata/telegraf/metric"
-	"github.com/influxdata/telegraf/plugins/outputs/wavefront"
 )
 
 func TestBuildTags(t *testing.T) {
@@ -106,11 +105,11 @@ func TestBuildTagsHostTag(t *testing.T) {
 
 func TestFormatMetricPoint(t *testing.T) {
 	var pointTests = []struct {
-		ptIn *wavefront.MetricPoint
+		ptIn *MetricPoint
 		out  string
 	}{
 		{
-			&wavefront.MetricPoint{
+			&MetricPoint{
 				Metric:    "cpu.idle",
 				Value:     1,
 				Timestamp: 1554172967,
@@ -120,7 +119,7 @@ func TestFormatMetricPoint(t *testing.T) {
 			"\"cpu.idle\" 1.000000 1554172967 source=\"testHost\" \"aaa\"=\"bbb\"\n",
 		},
 		{
-			&wavefront.MetricPoint{
+			&MetricPoint{
 				Metric:    "cpu.idle",
 				Value:     1,
 				Timestamp: 1554172967,
@@ -144,11 +143,11 @@ func TestFormatMetricPoint(t *testing.T) {
 
 func TestUseStrict(t *testing.T) {
 	var pointTests = []struct {
-		ptIn *wavefront.MetricPoint
+		ptIn *MetricPoint
 		out  string
 	}{
 		{
-			&wavefront.MetricPoint{
+			&MetricPoint{
 				Metric:    "cpu.idle",
 				Value:     1,
 				Timestamp: 1554172967,
